Inclusão no debug de diversas informações

+++ src/api/Request.php
@@ -31,13 +31,6 @@ class Request
 
     public function auth(Santander $credentials)
     {
-        #verifico se existe o certificado no servidor
-        if(!file_exists($credentials->getCertificate())){
-            throw new Exception("[ERRO] Não localizei o certificado no servidor!");
-        }
-        if(!file_exists($credentials->getCertificateKey())){
-            throw new Exception("[ERRO] Não localizei a chave do certificado no servidor!");
-        }
         #Inicio a autenticação
         $endpoint = $credentials->getEnvironment()->getApiUrl().'/auth/oauth/v2/token';
         $headers = [
@@ -68,6 +61,14 @@ class Request
         ]);
 
         try {
+            echo "Validando o certificado...<br>";
+            #verifico se existe o certificado no servidor
+            if(!file_exists($credentials->getCertificate())){
+                throw new Exception("[ERRO] Não localizei o certificado no servidor!");
+            }
+            if(!file_exists($credentials->getCertificateKey())){
+                throw new Exception("[ERRO] Não localizei a chave do certificado no servidor!");
+            }
             echo 'Vai executar...<br>';
             $response = curl_exec($curl);
             var_dump(['auth' => $response]);
